api/requirements.txt
uvicorn==0.23.2
fastapi==0.101.0
pytest==7.4.0
httpx==0.24.1
python-decouple==3.8
requests==2.31.0